Add the testing funtion
